UnityNative  IOSVolumeControllerErrorFix

+++ Unity/UnityDemo/Assets/Plugins/GF.Native/IOS/Native/AudioControl/VolumeController.h
@@ -20,6 +20,6 @@
 -(void) setMusicMax;
 -(void) cancelMusicMax;
 -(bool) getIsMaxVolume;
--(void) setCurrentMusicVolume:(int) current_muiscvolume;
+-(void) setCurrentMusicVolume:(float) current_muiscvolume;
 -(float) getCurrentMusicVolume;
 @end
